requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.10
aiosignal==1.3.2
attrs==25.3.0
beautifulsoup4==4.13.4
certifi==2025.4.26
requests==2.32.3
python-dotenv
google-cloud-storage
apache-airflow==3.0.1
psycopg2-binary
asyncpg
apache-airflow-providers-fab
apache-airflow-providers-celery
numpy==1.26.4
